+++ software/xbeerconfig.cpp
@@ -196,7 +196,7 @@ namespace {
 			std::cerr << "No such robot!\n";
 			return 1;
 		}
-		const Config::RobotInfo &botinfo(conf.robots().find(robot));
+		const Config::RobotInfo &botinfo(conf.robots().find(static_cast<unsigned int>(robot)));
 		XBeeLowLevel modem;
 
 		try {
